requirements.txt
# Core ML dependencies
torch>=1.9.0
torchvision>=0.10.0
open-clip-torch>=2.20.0
lightgbm>=3.3.0
scikit-learn>=1.0.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0
Pillow>=8.3.0
opencv-python>=4.5.0

# API and async processing
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
ray>=2.0.0

# Image processing and augmentation
albumentations>=1.1.0
albumentations-pytorch>=0.1.0

# MLflow for experiment tracking
mlflow>=1.20.0

# Visualization and plotting
matplotlib>=3.4.0
seaborn>=0.11.0

# Async HTTP client for load testing
aiohttp>=3.8.0

# Utilities
tqdm>=4.62.0

# File handling
openpyxl>=3.0.0

# System monitoring
psutil>=5.8.0

# Jupyter support (for notebooks)
ipywidgets>=7.6.0
jupyter>=1.0.0 